fix: 64-bit IPC ids, guard stoull, null-check malloc
- CEF carried call/callback/eval ids through CefValue SetInt/GetInt,
  truncating 64-bit ids to int32 across the renderer<->browser boundary.
  Carry them as double (exact to 2^53) at all 12 send/recv sites.
- webview JSON parser: guard std::stoull on __callback__ ids so a
  malformed payload returns null instead of throwing out of the parser.
- ValueDictKeys (CEF + webview): null-check malloc and roll back partial
  allocations instead of dereferencing null / leaking on OOM.

+++ cef/src/app.cc
@@ -330,7 +330,7 @@ bool WefHandler::OnProcessMessageReceived(
 
   if (name == "wef_call") {
     CefRefPtr<CefListValue> args = message->GetArgumentList();
-    uint64_t call_id = static_cast<uint64_t>(args->GetInt(0));
+    uint64_t call_id = static_cast<uint64_t>(args->GetDouble(0));
     std::string method_path = args->GetString(1).ToString();
     CefRefPtr<CefListValue> callArgs = args->GetList(2);
 
@@ -341,7 +341,7 @@ bool WefHandler::OnProcessMessageReceived(
 
   if (name == "wef_eval_result") {
     CefRefPtr<CefListValue> args = message->GetArgumentList();
-    uint64_t eval_id = static_cast<uint64_t>(args->GetInt(0));
+    uint64_t eval_id = static_cast<uint64_t>(args->GetDouble(0));
     CefRefPtr<CefValue> result = args->GetValue(1);
     std::string error = args->GetString(2).ToString();
     RuntimeLoader::GetInstance()->HandleEvalResult(eval_id, result, error);
